pyproject.toml
[tool.poetry]
name = "fastapi-gcp"
version = "0.1.0"
description = ""
authors = ["GlennViroux <glenn.viroux@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
tortoise-orm = "^0.19.3"
pydantic = {extras = ["dotenv"], version = "^1.10.8"}
asyncpg = "^0.27.0"
aerich = "^0.7.1"



[tool.aerich]
tortoise_orm = "database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
